Perfumy added
This project is a static website titled "Perfumy - Perfume showcase Website." It serves as an online showcase for various perfumes, featuring a visually appealing layout with high-quality images and brief descriptions of each product. The site includes a header with navigation links, a search bar for easy access, and sections dedicated to displaying perfumes, information about the site, and contact details. The design is clean and user-friendly, making it easy for visitors to browse through the showcased perfumes.